Fixed volume slider and ship spawning
Previously tried adding a collision mesh to the ships to avoid stacking, but it caused the ships to slide to the right when they spawned. This issue has been resolved.

The volume slider now sets to a minimum to -60 dB so the audio only goes silent when set to the minimum. Also set the 'step' variable to 1 so that the slider could be more precise.